fix build warnings: avoid narrowing conversion

+++ Source/ui/TabView.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2024 Marco Wang <m.aesophor@gmail.com>. All rights reserved.
+// Copyright (c) 2018-2025 Marco Wang <m.aesophor@gmail.com>. All rights reserved.
 
 #include "TabView.h"
 
@@ -42,7 +42,7 @@ void TabView::selectTab(int index) {
 void TabView::selectPrev() {
   int prev = _current - 1;
   if (prev < 0) {
-    prev = _tabs.size() - 1;
+    prev = static_cast<int>(_tabs.size()) - 1;
   }
   selectTab(prev);
 }
